aider: Updated NPC to always face the player character (Knight) as it moves.
# Aider chat conversation:

USER: where is the part that represents the current coordinates of the player (controlled by joystick movement)?

ASSISTANT: The current coordinates of the player, as controlled by joystick movement, are represented in the `Player.tsx` file within the `useFrame` hook. The `useFrame` hook updates the player's position based on the joystick's input, which is processed and applied as velocity to the player's physics body.

Here is the relevant section of the `Player.tsx` file that handles the player's position update:

```typescript
useFrame(({ gl, scene, camera }, delta) => {
  if (!ref.current) return;

  // ... (other code)

  const { x, z: y } = ref.current.position;

  tempVec2.set(x, y);
  largeColliderApi.setPosition(tempVec2);

  // ... (other code related to movement and velocity)

  playerPosition.x = x;
  playerPosition.y = y;
  playerPosition.angle = ref.current.rotation.y;

  // ... (other code)

  gl.render(scene, camera);
}, 100);
```

In this code, `playerPosition.x` and `playerPosition.y` are updated with the current position of the player, which is obtained from `ref.current.position`. The `tempVec2.set(x, y)` call is used to update the position of the player's large collider to match the player's position. The `useFrame` hook is called every frame, and thus the player's position is continuously updated based on the joystick input and the resulting velocity applied to the player's physics body.

USER: make the NPC always face the player character (Knight) as it moves.

ASSISTANT: To make the NPC always face the player character, we need to update the NPC's rotation to face the player's current position. This can be done by calculating the angle between the NPC and the player and then setting the NPC's rotation accordingly. We will use the `useFrame` hook from `@react-three/fiber` to continuously update the NPC's rotation every frame.

Here's the diff for the changes to the `NPC.tsx` file:

```diff
--- game/src/client/3d/models/Knight/NPC.tsx
+++ game/src/client/3d/models/Knight/NPC.tsx
@@ ... @@
import { useOutput } from "@livestack/client/src";
import { LiveJobContext } from "../../../game/components/Game/LiveJob";
+import { useFrame } from "@react-three/fiber";
+import { Vector3 } from "three";
import React from "react";
import { z } from "zod";
export function NPC({ ...props }: GroupProps) {
+  const playerPosition = useSnapshot(playerPosition);
+  const npcRef = useRef<GroupProps>(null);
+
+  useFrame(() => {
+    if (npcRef.current && playerPosition) {
+      const playerVec = new Vector3(playerPosition.x, npcRef.current.position.y, playerPosition.y);
+      npcRef.current.lookAt(playerVec);
+    }
+  });
+
  const position = [0, 0, 3] as [number, number, number];
  const job1 = React.useContext(LiveJobContext).conersationJob;
  if (!job1) return null;
@@ ... @@
  return (
-    <group position={position} rotation={[0, Math.PI, 0]} {...props}>
+    <group ref={npcRef} position={position} {...props}>
      {/* <Hud> */}
      <Html
        position={[-1, 9, 0]}
        prepend
        zIndexRange={[100, 0]}
        // position={[position[0], position[1] + 1, position[2]]}
      >
@@ ... @@
      </Html>
      <NPCCharacter
        lastAttack={0}
        lastDamaged={0}
        moving={false}
        recharging={false}
        running={false}
      />
    </group>
  );
}
```

In this diff, we've added a `useFrame` hook to the `NPC` component that updates the NPC's rotation to face the player's position every frame. We've also added a `ref` to the NPC's group element to be able to access and modify its properties. The `playerPosition` is obtained from the global state using `useSnapshot`, which should be imported from `valtio`. The `Vector3` class from `three` is used to create a vector representing the player's position, which is then used in the `lookAt` method to orient the NPC towards the player.