refactor: Update livestack submodule commit reference
This commit updates the reference to the submodule commit in the livestack file. The previous commit reference was 972eea41da6bb00f8fc0f8779ce0e39f99bb7752, and it has been updated to f36fe0ccdc83755ae70848f2762ba8143227f8a6. This change ensures that the correct version of the submodule is being used.

Co-authored-by: Xiyang Chen <settinghead@users.noreply.github.com>